docker/grpc-service/src/requirements.txt
requests
grpcio